+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/persistence/mapper/DbRecipeMapper.kt
@@ -14,7 +14,7 @@ object DbRecipeMapper {
             carbohydrates = recipe.carbohydrates,
             protein = recipe.protein,
             fat = recipe.fat,
-            servings = recipe.servings.toDouble(),
+            servings = recipe.servings,
             instructions = recipe.instructions,
             visibility = recipe.visibility,
             deleted = delete
@@ -30,7 +30,7 @@ object DbRecipeMapper {
             carbohydrates = recipeEntity.carbohydrates,
             protein = recipeEntity.protein,
             fat = recipeEntity.fat,
-            servings = recipeEntity.servings.toInt(),
+            servings = recipeEntity.servings,
             instructions = recipeEntity.instructions,
             visibility = recipeEntity.visibility,
             ingredients = ingredientEntities.map { ingredientWithFoodProduct ->
